app/requirements.txt
pytz
datetime
PyYaml
azure-iothub-device-client
w1thermsensor